pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memoryweb"
version = "0.1.0"
description = "Provenance-Aware Tiered Memory System"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "bcrypt>=4.1.0",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "sentence-transformers>=3.3.0",
    "numpy>=1.26.0",
    "python-multipart>=0.0.20",
    "PyJWT>=2.8.0",
    "aiofiles>=24.1.0",
    "python-dateutil>=2.9.0",
    "xxhash>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
tmp_path_retention_policy = "none"
# Use D: drive for temp to avoid AppData permission issues on Windows
addopts = ["--basetemp=D:/tmp/pytest"]
